requirements/base.txt
# Common base requirements shared by all environments

psycopg2==2.7.3.2
bcrypt==3.1.4

# ??? do we need it with webpack?
django-compressor==2.2
django-model-utils==3.0.0
django-forms-bootstrap==3.1.0
Django==2.0.1
pytz
raven==6.4.0
django-jsonify==0.3.0

django-settings-context-processor==0.2
django-admin-decorators==0.1
django-webpack-loader==0.5.0

django-cors-headers==2.2.0

# Django REST framework
djangorestframework==3.7.7
#drf-extensions==0.3.1 # NOTE: does not work with django 2.0 as of now

django-redis==4.8.0
redis==2.10.6

# celery related packages
celery==4.1.0
django-celery==3.2.2
django-celery-beat==1.4.0
flower==0.9.2

# Have to keep it at this version due to a bug in 4.1.0
kombu==4.0.2
#django-environ

# authentication
django-allauth==0.34.0
djangorestframework-jwt==1.11.0
django-rest-auth[with_social]==0.9.2

ua-parser==0.7.3
user-agents==1.1.0
django-user-agents==0.3.2

git+https://github.com/dj-stripe/dj-stripe@9145410fb4b067e8aad7f181ddc8b339a3514774#egg=dj_stripe
stripe==1.79.0
elasticsearch==6.1.1
django-rest-elasticsearch==0.3.4
elasticsearch-dsl==6.1.0

Pillow==5.0.0
requests==2.18.4
arrow==0.12.1

# scraping products
beautifulsoup4==4.6.0
ebaysdk==2.1.5

tldextract==2.2.0

git+https://gitlab.com/fpghost/python-skimlinks.git@master#egg=python_skimlinks

django-admin-sortable2==0.6.19

# Documentation
django-rest-swagger==2.2.0

Scrapy==1.5.1
scrapy-fake-useragent==1.1.0
scrapy-djangoitem==1.1.1
boto3==1.7.84

django-storages==1.7.1

# Video duration
moviepy==0.2.3.5

django-countries==5.3.2
country-currencies==0.2
retrying==1.3.3

# Segment
analytics-python==1.2.9

# Sendgrid 
sendgrid==5.6.0

git+https://github.com/monester/glacier-upload@master#egg=glacier_upload
